feat: Extraction de contrat pour faciliter la lisibilité du code
